Issue #27 - Fix CoreBlockContent to use already existing methods
